Rename compiler_plugins -> compiler/plugin/
